// src/modules/app/apps/dto/create-app.dto.ts
export class CreateAppDto {
  bussinessName: string;
  industryId: string;
  industryOther: string;
  website: string;
  mgpCommission: number;
  logo: string;
  showLogo: boolean;
  appName: string;
  appIcon: string;
  firebaseAppId: string;
  appBanner: string;
  socialMedia: string;
  socialMediaHandler: string;
  checklist: string;
  iosAppLink: string;
  andriodAppLink: string;
  webAppLink: string;
  userId: string;
}